+++ inc/templates.php
@@ -72,6 +72,24 @@ function mediatheque_print_containers( $editor = true ) {
 }
 
 /**
+ * Loads the wpview & the paste TinyMCE plugins when MediaThèque is used on front-end.
+ *
+ * The paste plugin is required as the wpview one is using it.
+ *
+ * @since 1.2.2
+ *
+ * @param  array $tinymce_plugins The tiny or Teeny MCE plugins.
+ * @return array                  The tiny or Teeny MCE plugins.
+ */
+function mediatheque_load_tinymce_plugins( $tinymce_plugins = array() ) {
+	if ( ! in_array( 'wpview', $tinymce_plugins, true ) ) {
+		$tinymce_plugins = array_merge( $tinymce_plugins, array( 'wpview', 'paste' ) );
+	}
+
+	return $tinymce_plugins;
+}
+
+/**
  * Outputs the Mediatheque button & UI.
  *
  * @since 1.0.0
@@ -119,7 +137,11 @@ function mediatheque_button( $args = array() ) {
 
 	if ( ! is_admin() ) {
 		wp_enqueue_style( 'mediatheque-front' );
+
+		// These scripts make sure the media are rendered on front-end.
 		wp_enqueue_script( 'mce-view' );
+		add_filter( 'tiny_mce_plugins',  'mediatheque_load_tinymce_plugins' );
+		add_filter( 'teeny_mce_plugins', 'mediatheque_load_tinymce_plugins' );
 	}
 
 	$img = '';